fix type error

+++ server.py
@@ -25,7 +25,7 @@ def receive_alert():
         search_name="avgReceive"
         #print(search_name)
         rx_bw=fetch_saved_search_results(saved_search_name=search_name)
-        remaining_bw=10000-int(rx_bw)
+        remaining_bw=10000-float(rx_bw)
         data_bw=10-int(totalParticipant)*1
         print(f'downstream used bandwidth: {rx_bw["avgRx"]}kbps')
         print(f'required bandwidth for Webex: {int(totalParticipant)*1}000kbps')
